requirements.txt
aiohttp==3.9.5
pandas==2.2.2
pandas-ta==0.3.14b0
numpy==1.26.4
scikit-learn==1.5.0
optuna==3.6.1
python-dotenv==1.0.1
discord-webhook==1.3.0
joblib==1.4.2
